Récupération des données depuis le site

+++ project/data/pull.py
@@ -1,27 +1,55 @@
 import csv
 from ifap.models import Classe, Ordre, Famille, Animal
 from math import floor
+from bs4 import BeautifulSoup
+import requests
+import re
+import pandas
+import io
 
-print("Pulling data...", end='\r')
 
-Classe.objects.all().delete()
+page = requests.get("https://www.i-fap.fr/espace-grand-public")
 
-with open('data/data.csv', newline='') as fichier:
-    data = csv.reader(fichier, delimiter=';')
-    table = list(enumerate((row for row in data), -1))
-    size = len(table) - 1
+print("Fetching data...")
 
-    for index, row in table[1:]:
-        print('Pulling data... {}%'.format(floor(index / size * 100)), end='\r')
+if page.status_code == 200:
+    pattern = re.compile("^https?://www\.i-fap\.fr/uploads/TAXO-i-fap-Grand-Public-V[0-9]+\.xlsx$")
 
-        classe = Classe.objects.get_or_create(nom=row[0])[0]
+    bsData = BeautifulSoup(page.content, "html.parser")
+    for a in bsData.find_all('a', href=True):
+        href = a.get('href')
+        if pattern.fullmatch(href) is not None:
+            try:
+                buffer = io.StringIO()
+                excelFile = pandas.read_excel(href)
+                excelFile.to_csv(buffer, ';', index=False, encoding="utf-8")
+                buffer.seek(0)
 
-        ordre = Ordre.objects.get_or_create(nom=row[1], classe=classe)[0]
+                print("Merging data...", end='\r')
 
-        famille = Famille.objects.get_or_create(nom=row[2], ordre=ordre)[0]
+                Classe.objects.all().delete()
 
-        Animal.objects.get_or_create(nom_scientifique=row[3], nom_commun=row[4], famille=famille)
+                data = csv.reader(buffer, delimiter=';')
+                table = list(enumerate((row for row in data)))
+                size = len(table) - 1
 
-    print("Data have been pulled with success")
+                for index, row in table[1:]:
 
-    fichier.close()
+                    classe = Classe.objects.get_or_create(nom=row[0])[0]
+
+                    ordre = Ordre.objects.get_or_create(nom=row[1], classe=classe)[0]
+
+                    famille = Famille.objects.get_or_create(nom=row[2], ordre=ordre)[0]
+
+                    Animal.objects.get_or_create(nom_scientifique=row[3], nom_commun=row[4], famille=famille)
+
+                    print('Merging data... {}%'.format(floor(index / size * 100)), end='\r')
+
+                print("\nData have been pulled with success")
+                buffer.close()
+
+            except:
+                print("Error on merge data")
+
+else:
+    print("Error on fetch data")
